+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log (vs-deploy)
 
+## 9.11.0 (May 29th, 2017; settings)
+
+* added `alwaysShowPackageList` setting that indicates if package list is shown, even if there is only 1 entry
+* added `alwaysShowTargetList` setting that indicates if target list is shown, even if there is only 1 entry
+
 ## 9.10.0 (May 28th, 2017; new list target)
 
 * added [list target](https://github.com/mkloubert/vs-deploy/wiki/target_list), that lets the user select an entry by the GUI, which contains settings for one or more other target